backend/requirements.txt
dotenv
fastapi
uvicorn
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]

# Dependências de teste
pytest
httpx